move xml test util code
