+++ corehq/tests/util/xml.py
@@ -12,6 +12,17 @@ def assert_xml_equal(expected, actual, normalize=True):
     _check_shared(expected, actual, LXMLOutputChecker(), "xml")
 
 
+def assert_xml_partial_equal(expected, actual, xpath):
+    """
+    Extracts a section of XML using the xpath and compares it to the expected
+
+    Extracted XML is placed inside a <partial/> element prior to comparison.
+    """
+    expected = parse_normalize(expected)
+    actual = extract_xml_partial(actual, xpath)
+    assert_xml_equal(expected, actual, normalize=False)
+
+
 def assert_html_equal(expected, actual, normalize=True):
     if normalize:
         expected = parse_normalize(expected, is_html=True)
@@ -62,3 +73,12 @@ def _check_shared(expected, actual, checker, extension):
             # check that there was actually a diff, because checker.check_output
             # doesn't work with unicode characters in xml node names
             raise AssertionError(message)
+
+
+def extract_xml_partial(xml, xpath):
+    actual = parse_normalize(xml, to_string=False)
+    nodes = actual.findall(xpath)
+    root = lxml.etree.Element('partial')
+    for node in nodes:
+        root.append(node)
+    return lxml.etree.tostring(root, pretty_print=True, encoding='utf-8')
